backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
anthropic>=0.18.0
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
httpx>=0.25.0
selenium==4.15.0
webdriver-manager==4.0.1
python-socketio==5.10.0
flask-socketio==5.3.5
websockets==12.0
pillow==10.1.0

